requirements.txt
Flask~=3.0.1
Bootstrap_Flask~=2.3.3
requests~=2.31.0
WTForms~=3.1.2
Flask-WTF~=1.2.1
email-validator~=2.1.0.post1
python-dotenv~=1.0.1